global styling for some pages, global font, and moving files to src directory
