--- components/Card/Card.module.css
@@ -1,17 +1,0 @@
-.card {
-    width: 100px;
-    height: 150px;
-    position: absolute;
-    transition: transform 0.3s, filter 0.3s;
-    cursor: pointer;
-    filter: none; /* Ensure full color */
-  }
-  
-  .card:hover {
-    transform: scale(1.1);
-    z-index: 1;
-  }
-  
-  .hiddenCard {
-    display: none;
-  }
