--- pages/about.tsx
@@ -1,3 +1,0 @@
-export default function About() {
-  return <h1>About Page</h1>;
-}
