--- pages/identity.tsx
@@ -1,3 +1,0 @@
-export default function Identity() {
-  return <h1>Me</h1>;
-}
