+++ src/pages/about.tsx
@@ -1,0 +1,10 @@
+import Navbar from "./components/NavBar/NavBar";
+
+export default function About() {
+  return (
+    <>
+      <Navbar />
+      <h1>About Page</h1>
+    </>
+  );
+}
