+++ src/pages/identity.tsx
@@ -1,0 +1,10 @@
+import Navbar from "./components/NavBar/NavBar";
+
+export default function Identity() {
+  return (
+    <div>
+      <Navbar />
+      <h1>Me</h1>
+    </div>
+  );
+}
