+++ src/pages/styles/styles.css
@@ -1,8 +1,3 @@
-@font-face {
-  font-family: "Mermaid";
-  src: local("../pages/mermaid/Mermaid1001.ttf");
-}
-
 * {
   --color-text-base: #000;
   --color-fill-home: #ff005c;
@@ -23,23 +18,6 @@
   box-sizing: border-box;
 }
 
-body {
-}
-
-button {
-  font-family: "Mermaid";
-  font-weight: 700;
-}
-
 div > p {
   margin: 1rem;
 }
-
-/* .container {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  width: 100vw;
-  align-items: center;
-  justify-content: center;
-} */
