--- styles/CardReveal.module.css
@@ -1,48 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  background-color: var(--color-fill-card-reveal);
-}
-
-.card {
-  width: 300px;
-  height: auto;
-  margin-bottom: 20px;
-  z-index: 2;
-}
-
-.descriptionBox {
-  margin-top: 20px;
-  padding: 10px;
-  background-color: var(--color-fill-container);
-  border: 3px solid black;
-  border-radius: 5px;
-  color: black;
-  text-align: center;
-}
-
-.button {
-  padding: 10px 20px;
-  background-color: var(--color-fill-lavender);
-  color: black;
-  border-radius: 5px;
-  border: 3px solid black;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  margin-top: 1rem;
-  font-size: 1.25rem;
-}
-
-.button:hover {
-  background-color: var(--color-fill-deep-purple);
-}
-
-.error {
-  color: red;
-  font-weight: bold;
-  text-align: center;
-  margin-top: 50px;
-}
