--- styles/CardSelect.module.css
@@ -1,128 +1,0 @@
-.container {
-  display: flex;
-  background-color: var(--color-fill-card-select);
-  flex-direction: column;
-  height: 100vh;
-  width: 100vw;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  background-color: var(--color-fill-container);
-  padding: 10px;
-  border-bottom: 3px solid black;
-}
-
-.headerLink {
-  color: black;
-  font-weight: bold;
-  text-decoration: none;
-  margin: 0 10px;
-  text-transform: uppercase;
-}
-
-.main {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-grow: 1;
-  position: relative;
-  border-radius: 50%;
-  min-height: 50rem;
-}
-
-.cardDeck {
-  position: relative;
-  height: 600px;
-  width: 600px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.cardDeckDark {
-  position: relative;
-  height: 600px;
-  width: 600px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  filter: brightness(0.5); /* Darken the card deck */
-}
-
-.card {
-  width: 100px;
-  height: 150px;
-  position: absolute;
-  transition: transform 0.3s, filter 0.3s;
-  cursor: pointer;
-  filter: none; /* Ensure full color */
-}
-
-.card:hover {
-  transform: scale(1.1);
-  z-index: 1;
-}
-
-.hiddenCard {
-  display: none;
-}
-
-.centerText {
-  position: absolute;
-  text-align: center;
-  color: black;
-  font-weight: bold;
-  text-transform: uppercase;
-  line-height: 2rem;
-  font-size: 1.5rem; 
-  text-wrap: wrap;
-  width: 40%;
-  min-width: 20rem;
-}
-
-.selectedCardContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  z-index: 2; /* Ensure the selected card is above the darkened deck */
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.mintedCard {
-  width: 300px;
-  height: auto;
-}
-
-.selectedCardContainer .buttons {
-  display: flex;
-  gap: 20px;
-  margin-top: 20px;
-}
-
-.button {
-  padding: 10px 20px;
-  background-color: var(--color-fill-container);
-  color: black;
-  border-radius: 5px;
-  border: 3px solid black;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.button:hover {
-  background-color: grey;
-}
-
-.descriptionBox {
-  margin-top: 20px;
-  padding: 10px;
-  background-color: white;
-  border: 1px solid black;
-  color: black;
-}
