--- styles/Home.module.css
@@ -1,79 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  width: 100vw;
-  background-color: theme("colors.neonMagenta");
-  margin: 0;
-  position: relative;
-}
-
-.redBorder {
-  border: 5px solid black;
-  box-sizing: border-box;
-  height: 90vh;
-  width: 90vw;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.yellowFill {
-  background-color: var(--color-fill-container);
-  height: 100%;
-  width: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.archContainer {
-  position: relative;
-  height: 200px;
-  width: 100px;
-  margin-bottom: 20px;
-}
-
-/*going to revisit naming conventions, left for the time being*/
-
-.blueArch {
-  position: absolute;
-  width: 100px;
-  height: 200px;
-  background-color: var(--color-fill-deep-purple);
-  border-top-left-radius: 50px;
-  border-top-right-radius: 50px;
-  border: 2px solid black;
-}
-
-.pinkArch {
-  position: absolute;
-  width: 100px;
-  height: 200px;
-  background-color: var(--color-fill-lavender);
-  border-top-left-radius: 50px;
-  border-top-right-radius: 50px;
-  border: 2px solid black;
-  left: 10px;
-  top: 0;
-  transition: background-color 0.3s;
-}
-
-.pinkArch:hover {
-  background-color: yellow; /* Yellow on hover */
-}
-
-.enter {
-  font-size: 2em;
-  font-weight: bold;
-  color: black;
-  text-decoration: none;
-  transition: color 0.3s;
-}
-
-.enter:hover {
-  color: pink; /* Pink on hover */
-}
